// types.ts
export interface ServiceInfo {
  id: string;
  title: string;
  shortTitle: string;
  description: string;
  gymApplication: string;
  companyApplication: string;
  imageUrl: string;
  relatedIds: string[];
}

export interface HomeBanner {
  title: string;
  subtitle: string;
  imageUrl: string;
  cta: string;
}